Add mProcessor.join() to impl/PeerConnection::close(). Reset impl/PeerConnection dtor to default
This fixes a race condition that will occur when the shared_pointer of the impl() PeerConnection has more use_count than 1 at the time of the destruction of the wrapping PeerConnection, causing callbacks to still fire after the wrapper is dropped, which in turn is due to the fact that the impl() PeerConnection's destructor is never called. Since the wrapping PeerConnection is the only way to initialize the impl() PeerConnection, it is safe to only join in close(), which is always called upon wrapping PeerConnection dtor.