Add a media handler to respond to intra frame requests
This introduces a PLIResponder class, which can be used for responding
back with an intra frame after the receiver requests one with an FIR
(Full Intra Request) or a PLI (Picture Loss Indicator) message.

This is useful when a video is being streamed live and we can control
the behavior of the video encoder. PLIResponder simply notifies its
caller when a new intra frame is requested, which passes this request
onto its video frame source.