feat: update current remote description with fingerprint
Adds a `currentRemoteDescription` field to `PeerConnection` which
can be updated independently of `remoteDescription`, the idea being
the `remoteDescription` is the offer received from the remote peer
and the `currentRemoteDescription` is what has actually been
negotiated with them.

The `currentRemoteDescription` is initialised when the
`remoteDescription` is set.

Currently the `checkFingerprint` method updates the certificate
fingerprint used by the connection to the remote peer but I guess
other modifications may be useful.

Exposes the `currentRemoteDescription` via a getter similar to
`remoteDescription`.